+++ packages/hono/serices/user.ts
@@ -4,9 +4,16 @@ import { BaseService } from "./base-service";
 export class UserService extends BaseService {
   async createUser(values: Prisma.UserCreateInput) {
     try {
+      const existing = await this.getUserByEmail(values.email);
+
+      if (existing) {
+        throw new Error(`User already exists with this email.`);
+      }
+
       const newUser = await prisma.user.create({
         data: values,
       });
+
       return newUser;
     } catch (error) {
       throw new Error(`Error creating user: ${(error as Error).message}`);
@@ -15,6 +22,12 @@ export class UserService extends BaseService {
 
   async deleteUser(id: string) {
     try {
+      const exists = await this.getUserById(id);
+
+      if (!exists) {
+        throw new Error(`users does not exists with this id.`);
+      }
+
       await prisma.user.delete({
         where: { clerkId: id },
       });
@@ -26,6 +39,12 @@ export class UserService extends BaseService {
 
   async updateUser(id: string, values: Prisma.UserUpdateInput) {
     try {
+      const exists = await this.getUserById(id);
+
+      if (!exists) {
+        throw new Error(`users does not exists with this id.`);
+      }
+
       const updatedUser = await prisma.user.update({
         where: { clerkId: id },
         data: values,
@@ -40,6 +59,19 @@ export class UserService extends BaseService {
     try {
       const user = await prisma.user.findUnique({
         where: { clerkId: id },
+      });
+      if (!user) {
+        throw new Error("user not found");
+      }
+      return user;
+    } catch (error) {
+      throw new Error(`Error retrieving user: ${(error as Error).message}`);
+    }
+  }
+  async getUserByEmail(email: string) {
+    try {
+      const user = await prisma.user.findUnique({
+        where: { email: email },
       });
       if (!user) {
         throw new Error("user not found");
